+++ Start-IntuneManagement.ps1
@@ -1,7 +1,11 @@
 [CmdletBinding(SupportsShouldProcess=$True)]
 param(
     [switch]
-    $ShowConsoleWindow
+    $ShowConsoleWindow,
+    [switch]
+    $JSonSettings,
+    [string]
+    $JSonFile
 )
 Import-Module ($PSScriptRoot + "\CloudAPIPowerShellManagement.psd1") -Force
-Initialize-CloudAPIManagement -View "IntuneGraphAPI" -ShowConsoleWindow:($ShowConsoleWindow)
+Initialize-CloudAPIManagement -View "IntuneGraphAPI" -ShowConsoleWindow:($ShowConsoleWindow) -JSonSettings:($JSonSettings) -JSonFile $JSonFile